Time: 0 ms (100.00%) | Memory: 32.1 MB (96.03%) - LeetSync

+++ 4214-count-dominant-indices/count-dominant-indices.cpp
@@ -9,7 +9,7 @@ public:
         for(int i = 0; i < n - 1; i++){
             totalSum -= nums[i];
 
-            double avg = totalSum / deno;
+            double avg = (double)totalSum / deno;
             deno--;
             if(nums[i] > avg){
                 count++;
